fix compiler optimizations issue and finished program

+++ BigTacToe/Board.cpp
@@ -2,13 +2,13 @@
 
 #include <iostream>
 
-Board::Board(unsigned short int boardSize)
+Board::Board(unsigned short boardSize)
 {
 	size = boardSize;
 	states.resize(size, std::vector<States>(size));
 }
 
-unsigned short int Board::getBoardSize() const
+unsigned short Board::getBoardSize() const
 {
 	return size;
 }
@@ -16,18 +16,18 @@ unsigned short int Board::getBoardSize() const
 void Board::printBoard() const
 {
 	std::cout << "    ";
-	for (int i = 0; i < states.size(); i++)
+	for (unsigned short i = 0; i < states.size(); i++)
 		std::cout << i + 1 << "   ";
 
 	std::cout << std::endl << "  ";
-	for (int i = 0; i < (states.size() * 4) + 1; i++)
+	for (unsigned short i = 0; i < (states.size() * 4) + 1; i++)
 		std::cout << "-";
 	std::cout << std::endl;
 
-	for (int i = 0; i < states.size(); i++)
+	for (unsigned short i = 0; i < states.size(); i++)
 	{
 		std::cout << i + 1 << " ";
-		for (int j = 0; j < states[i].size(); j++)
+		for (unsigned short j = 0; j < states[i].size(); j++)
 		{
 			std::cout << "|";
 			if (states[i][j] == States::X) std::cout << " X ";
@@ -37,14 +37,14 @@ void Board::printBoard() const
 		}
 		std::cout << "|" << std::endl;
 		std::cout << "  ";
-		for (int k = 0; k < (states.size() * 4) + 1; k++)
+		for (unsigned short k = 0; k < (states.size() * 4) + 1; k++)
 			std::cout << "-";
 		std::cout << std::endl;
 	}
 	std::cout << std::endl;
 }
 
-bool Board::placeMark(unsigned short int xPos, unsigned short int yPos, States state)
+bool Board::placeMark(unsigned short xPos, unsigned short yPos, States state)
 {
 	if (state == States::Nil) {
 		return false;
@@ -61,7 +61,7 @@ bool Board::placeMark(unsigned short int xPos, unsigned short int yPos, States s
 
 }
 
-States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) const
+States Board::checkBoardWin(unsigned short xPos, unsigned short yPos) const
 {
 	const States playerState = states[xPos - 1][yPos - 1];
 
@@ -69,7 +69,7 @@ States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) co
 		return States::Nil;
 
 	// check for column win
-	for (int i = 0; i < size; i++) {
+	for (unsigned short i = 0; i < size; i++) {
 		if (states[xPos - 1][i] != playerState)
 			break;
 		if (i == size - 1)
@@ -77,7 +77,7 @@ States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) co
 	}
 
 	// check for row win
-	for (int i = 0; i < size; i++) {
+	for (unsigned short i = 0; i < size; i++) {
 		if (states[i][yPos - 1] != playerState)
 			break;
 		if (i == size - 1)
@@ -86,7 +86,7 @@ States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) co
 
 	// check diagonal win
 	if (xPos - 1 == yPos - 1) {
-		for (int i = 0; i < size; i++) {
+		for (unsigned short i = 0; i < size; i++) {
 			if (states[i][i] != playerState)
 				break;
 			if (i == size - 1)
@@ -96,7 +96,7 @@ States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) co
 
 	// check anti diagonal win
 	if (xPos - 1 + yPos - 1 == size - 1) {
-		for (int i = 0; i < size; i++) {
+		for (unsigned short i = 0; i < size; i++) {
 			if (states[i][(size - 1) - i] != playerState)
 				break;
 			if (i == size - 1)
@@ -106,7 +106,7 @@ States Board::checkBoardWin(unsigned short int xPos, unsigned short int yPos) co
 
 	return States::Nil;
 }
-States Board::getState(unsigned short int xPos, unsigned short int yPos) const
+States Board::getState(unsigned short xPos, unsigned short yPos) const
 {
 	return states[xPos - 1][yPos - 1];
 }
